Update from .Net Framework to .Net Core
