+++ Models/Summery.cs
@@ -1,0 +1,8 @@
+﻿namespace SpawnCodeGenerator.Models
+{
+    public class Summery
+    {
+        public SummeryEnum Type { get; set; }
+        public string Value { get; set; }
+    }
+}
