+++ Models/SummeryEnum.cs
@@ -1,0 +1,11 @@
+﻿namespace SpawnCodeGenerator.Models
+{
+    public enum SummeryEnum
+    {
+        BLUEPRINT,
+        ENGRAM,
+        PRIMAL_ITEM,
+        SPAWNCODE_ITEM,
+        SPAWNCODE_CREATURE
+    }
+}
